+++ mtui/commands/update.py
@@ -80,7 +80,7 @@ class Update(Command):
             logger.critical("failed to update target systems")
             logger.debug(format_exc())
             self.prompt.notify_user(
-                f"updating {self.prompt.session!s} failed", "stock_dialog-error"
+                f"updating {self.metadata.id!s} failed", "stock_dialog-error"
             )
             raise
 
@@ -88,7 +88,7 @@ class Update(Command):
             logger.info("update process canceled")
             return
 
-        self.prompt.notify_user(f"updating {self.prompt.session!s} finished")
+        self.prompt.notify_user(f"updating {self.metadata.id!s} finished")
         logger.info("done")
 
     @staticmethod